requirements.txt
python-docx
python-hcl2
diagrams